+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+    "cSpell.words": [
+        "articlebrief",
+        "headerscrape",
+        "newtitlegen",
+        "outlinegen"
+    ]
+}
