+++ ui/src/modules/Titles.js
@@ -1,17 +1,16 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect } from 'react';
+import ProgressBar from '../lib/Progressbar';
 
-function StepFive({ nextStep, stepData }) {
+function Titles({ nextStep, stepData, nextModule }) {
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [selectedTitle, setSelectedTitle] = useState('');
   const [customTitle, setCustomTitle] = useState('');
   const [useCustomTitle, setUseCustomTitle] = useState(false);
-  const [progress, setProgress] = useState(0);
-  const [bars, setBars] = useState("░");
   const [errorMessage, setErrorMessage] = useState('');
-  const progressInterval = useRef(null);
 
   useEffect(() => {
     console.log('Titles Stepdata on load:', stepData);
+    console.log('Next Module:', nextModule);
   }, [stepData]);
 
   const handleTitleChange = (title) => {
@@ -29,32 +28,9 @@ function StepFive({ nextStep, stepData }) {
       return;
     }
     setIsSubmitting(true);
-    initProgressBar();
     const submissionData = useCustomTitle && customTitle ? { customTitle } : { selectedTitle };
     await nextStep({ ...stepData, ...submissionData });
   };
-
-  const initProgressBar = () => {
-    let intervalDuration = 200; // 50 seconds for demo purposes
-    progressInterval.current = setInterval(() => {
-      setProgress((prevProgress) => {
-        const newProgress = prevProgress < 100 ? prevProgress + 1 : 100;
-        if (newProgress === 100) {
-          clearInterval(progressInterval.current);
-          setBars("░".repeat(41)); // Assuming 41 bars represent 100%
-          console.log("progress complete");
-        } else if (newProgress % 5 === 0) {
-          setBars((prevBars) => prevBars + "░░");
-          console.log("progress", newProgress);
-        }
-        return newProgress;
-      });
-    }, intervalDuration);
-  };
-
-  useEffect(() => {
-    return () => clearInterval(progressInterval.current);
-  }, []);
 
   return (
     <div>
@@ -118,7 +94,7 @@ function StepFive({ nextStep, stepData }) {
 
       <div className="flex justify-end mt-4">
         <button
-          className={`px-4 py-2 ${isSubmitting ? 'bg-blue-300' : 'bg-blue-500 hover:bg-blue-700'} text-white rounded transition-colors`}
+          className={`px-4 py-2 ${isSubmitting ? 'bg-gray-500' : 'bg-blue-500 hover:bg-blue-700'} text-white rounded transition-colors`}
           onClick={handleSubmit}
           disabled={isSubmitting}
         >
@@ -130,23 +106,17 @@ function StepFive({ nextStep, stepData }) {
               </svg>
               Processing...
             </div>
-          ) : (
-            'Add Article Brief'
-          )}
+          ) : nextModule.buttonLabel ? nextModule.buttonLabel : 'Next'}
         </button>
       </div>
-      {isSubmitting && (
-        <>
-          <div className="mt-5">
-            <div className="progress-bar">
-              <div className="bg-blue-500 h-1" style={{ width: `${progress}%` }}></div>
-              {bars}{progress}%
-            </div>
-          </div>
-        </>
+      {isSubmitting && nextModule.hasProgressBar && (
+        <ProgressBar 
+          executionTime={nextModule.executionTime} 
+          renderProgressMessage={nextModule.renderProgressMessage} 
+        />
       )}
     </div>
   );
 }
 
-export default StepFive;
+export default Titles;
